requirements.txt
aiohttp==3.8.6
aiomqtt==1.2.1
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
lxml==4.9.3
multidict==6.0.4
paho-mqtt==1.6.1
pycparser==2.21
PyJWT==2.8.0
PyYAML==6.0.1
skodaconnect==1.3.8
sniffio==1.3.0
soupsieve==2.5
yarl==1.9.2
